Move settings handlers out of main
